+++ app/src/main/java/com/depromeet/sloth/ui/list/TodayLessonAdapter.kt
@@ -1,4 +1,4 @@
-package com.depromeet.sloth.ui.lesson
+package com.depromeet.sloth.ui.list
 
 import android.animation.ObjectAnimator
 import android.graphics.Color
